refactor(gate): extract the ONE gate engine — subject × prior × policy (4.4.0 WP1) (#291)
* refactor(gate): extract the ONE gate engine — subject × prior × policy (4.4.0 WP1)

runGuardrailCheck's orchestration moves to src/gate/engine.ts:runGate,
with the seams the release builds on split out per concept: prior
acquisition (prior.ts — Rule 11 applied to the engine; the ref arm and
the WP2 dir-shaped arms share scanToBaselineFile), pair classification
(classify-pairs.ts), observation semantics (observation.ts), per-pair
context helpers (context.ts), and the result/request vocabulary
(result.ts, types.ts). src/baseline/check.ts stays the guardrail
surface and the stable import path — every moved name re-exported
identically, all existing consumers untouched. Behavior frozen by the
WP0 parity net (verdicts, identities, disclosures, determinism all
pinned before the split). The arch-check's createHash / line-bucketing /
catch-swallow scopes extend to src/gate/ so the protective nets follow
the code.


* test(baseline): advisory-attribution source pin follows the consumers into src/gate/


---------
